requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2

# RAG & NLP
sentence-transformers==2.2.2
chromadb==0.4.18
openai==1.3.7

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
pyyaml==6.0.1
